+++ Omega/src/com/saggitt/omega/smartspace/OmegaSmartspaceController.kt
@@ -245,7 +245,7 @@ class OmegaSmartspaceController(val context: Context) {
             (Class.forName(className).getConstructor(OmegaSmartspaceController::class.java)
                     .newInstance(this) as DataProvider)
         } catch (t: Throwable) {
-            Log.d("LSC", "couldn't create provider", t)
+            Log.d("SmartspaceController", "couldn't create provider", t)
             BlankDataProvider(this)
         }
     }
@@ -545,7 +545,7 @@ class OmegaSmartspaceController(val context: Context) {
         private val displayNames = mapOf(
                 Pair(BlankDataProvider::class.java.name, R.string.weather_provider_disabled),
                 Pair(SmartspaceDataWidget::class.java.name, R.string.google_app),
-                //Pair(SmartspacePixelBridge::class.java.name, R.string.smartspace_provider_bridge),
+                Pair(SmartspacePixelBridge::class.java.name, R.string.smartspace_provider_bridge),
                 Pair(PEWeatherDataProvider::class.java.name, R.string.weather_provider_pe),
                 Pair(OnePlusWeatherDataProvider::class.java.name, R.string.weather_provider_oneplus_weather),
                 Pair(NowPlayingProvider::class.java.name, R.string.event_provider_now_playing),